# swagger.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/concert-dates/{concertDatesId}/seats": {
      "get": {
        "operationId": "SeatController_getAllSeats",
        "summary": "좌석 목록 조회",
        "parameters": [
          {
            "name": "concertDatesId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "participation-token",
            "in": "header",
            "description": "Participation token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Authorization token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Seat"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "seat"
        ]
      }
    }
  },
  "info": {
    "title": "Ticket Server API",
    "description": "Ticket Server API description",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SeatGrade": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "enum": [
              "S",
              "A",
              "B",
              "C"
            ]
          },
          "price": {
            "type": "number"
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "name",
          "price"
        ]
      },
      "Concert": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "format": "string"
          },
          "artist": {
            "type": "string",
            "format": "string"
          },
          "venue": {
            "type": "string",
            "format": "string"
          },
          "description": {
            "type": "string",
            "format": "string"
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "name",
          "artist",
          "venue",
          "description"
        ]
      },
      "ConcertDates": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "concertDate": {
            "format": "date-time",
            "type": "string"
          },
          "reserveDate": {
            "format": "date-time",
            "type": "string"
          },
          "concert": {
            "$ref": "#/components/schemas/Concert"
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "concertDate",
          "reserveDate",
          "concert"
        ]
      },
      "Seat": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "section": {
            "type": "string",
            "format": "string"
          },
          "row": {
            "type": "string",
            "format": "string"
          },
          "number": {
            "type": "number"
          },
          "isBooked": {
            "type": "boolean",
            "default": false
          },
          "seatGrade": {
            "$ref": "#/components/schemas/SeatGrade"
          },
          "concertDates": {
            "$ref": "#/components/schemas/ConcertDates"
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "section",
          "row",
          "number",
          "isBooked",
          "seatGrade",
          "concertDates"
        ]
      },
...
    }
  }
}